requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
idna==2.7
Markdown==2.6.11
PyGithub==1.40
PyJWT==1.6.4
requests==2.19.1
urllib3==1.23
